+++ packages/app-builder-lib/src/electron/electronVersion.ts
@@ -6,7 +6,7 @@ import { httpExecutor } from "builder-util"
 import { readJson } from "fs-extra"
 import { Lazy } from "lazy-val"
 import * as path from "path"
-import { orNullIfFileNotExist } from "../util/config/loadConfig"
+import { orNullIfFileNotExist } from "../util/config/load"
 import * as semver from "semver"
 import { Configuration } from "../configuration"
 import { getConfig } from "../util/config/config"
