+++ packages/electron-builder/src/cli/cli.ts
@@ -5,7 +5,7 @@ import * as chalk from "chalk"
 import { readJson } from "fs-extra"
 import * as isCi from "is-ci"
 import * as path from "path"
-import { loadEnv } from "app-builder-lib/out/util/config/loadConfig"
+import { loadEnv } from "app-builder-lib/out/util/config/load"
 import { build, configureBuildCommand, createYargs } from "../builder"
 import { createSelfSignedCert } from "./create-self-signed-cert"
 import { configureInstallAppDepsCommand, installAppDeps } from "./install-app-deps"
